Refactor controllers to split out functionality
