+++ app/views/returns/submittedReturns.scala.html
@@ -60,7 +60,7 @@
                 @inactiveTab(
                   messages("submittedReturns.previousReturns"),
                   messages("submittedReturns.inactivePrevious"),
-                  controllers.routes.ReturnObligationsController.submittedReturns(model.returnYears.last - 1).url
+                  controllers.routes.SubmittedReturnsController.submittedReturns(model.returnYears.last - 1).url
                 )
               }
             </ul>
