+++ app/views/templates/returns/submittedReturnsTabs.scala.html
@@ -25,7 +25,7 @@
     @inactiveTab(
       yearToDisplay.toString,
       messages("submittedReturns.inactiveTab", yearToDisplay.toString),
-      controllers.routes.ReturnObligationsController.submittedReturns(yearToDisplay).url
+      controllers.routes.SubmittedReturnsController.submittedReturns(yearToDisplay).url
     )
   }
 }
