+++ test/views/VatReturnDetailsViewSpec.scala
@@ -137,9 +137,9 @@ class VatReturnDetailsViewSpec extends ViewBaseSpec {
         elementText(Selectors.previousPageBreadcrumb) shouldBe "Submitted returns"
       }
 
-      s"link to ${controllers.routes.ReturnObligationsController.submittedReturns(currentYear).url}" in {
+      s"link to ${controllers.routes.SubmittedReturnsController.submittedReturns(currentYear).url}" in {
         element(Selectors.previousPageBreadcrumbLink).attr("href") shouldBe
-          controllers.routes.ReturnObligationsController.submittedReturns(currentYear).url
+          controllers.routes.SubmittedReturnsController.submittedReturns(currentYear).url
       }
 
       "have the correct current page text containing the obligation dates" in {
